rename bootstrap.rb -> preloads.rb and h2o.rb -> bootstrap.rb
