+++ share/h2o/mruby/acl.rb
@@ -18,7 +18,7 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 # IN THE SOFTWARE.
 #
-require "h2o.rb"
+require "bootstrap.rb"
 
 module H2O
