+++ share/h2o/mruby/preloads.rb
@@ -1,0 +1,3 @@
+require "bootstrap.rb"
+require "acl.rb"
+include H2O::ACL
